@@ -110,35 +111,52 @@
     "eslint-plugin-vue-storefront": "^0.0.1",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
-    "karma": "^1.1.2",
-    "karma-mocha": "^1.1.1",
+    "karma": "^1.7.1",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-mocha": "^1.3.0",
     "karma-phantomjs-launcher": "^1.0.1",
     "karma-sinon-chai": "^1.2.3",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^3.0.0",
     "lerna": "^2.9.0",
-    "markdown-to-vue-loader": "^0.4.0",
-    "mocha": "^3.0.2",
+    "markdown-to-vue-loader": "^0.5.0",
+    "mocha": "^3.5.3",
     "node-sass": "^4.9.0",
     "phantomjs-prebuilt": "^2.1.10",
+    "postcss-flexbugs-fixes": "^3.3.1",
+    "postcss-loader": "^2.1.5",
     "rimraf": "^2.6.0",
     "sass-loader": "^7.0.1",
+    "shelljs": "^0.8.1",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
-    "shelljs": "^0.8.1",
     "sw-precache-webpack-plugin": "^0.11.5",
     "url-loader": "^1.0.1",
-    "val-loader": "^1.1.0",
+    "val-loader": "^1.1.1",
     "vue-eslint-parser": "^2.0.3",
     "vue-loader": "^15.1.0",
     "vue-ssr-webpack-plugin": "^3.0.0",
     "vue-template-compiler": "^2.5.16",
-    "webpack": "^4.8.3",
+    "webpack": "^4.12.2",
     "webpack-cli": "^2.1.3",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.2",
     "webpack-merge": "^4.1.2"
+  },
+  "browserslist": {
+    "development": [
+      "last 2 chrome versions",
+      "last 2 firefox versions",
+      "last 2 edge versions",
+      "IE 10"
+    ],
+    "production": [
+      ">0.5%",
+      "last 4 versions",
+      "Firefox ESR",
+      "not ie < 10"
+    ]
   },
   "workspaces": [
     "core/store",
