@@ -1,4 +1,15 @@
 
 export default {
-  name: 'ForgotPass'
+  name: 'ForgotPass',
+  data () {
+    return {
+      email: '',
+      passwordSent: false
+    }
+  },
+  methods: {
+    switchElem () {
+      this.$store.commit('ui/setAuthElem', 'login')
+    }
+  }
 }
