@@ -1,3 +1,10 @@
+import config from 'config'
+
 export default {
-  name: 'MainFooter'
+  name: 'MainFooter',
+  computed: {
+    multistoreEnabled () {
+      return config.storeViews.multistore
+    }
+  }
 }
