@@ -1,4 +1,5 @@
-import { mapActions, mapState, mapGetters } from 'vuex'
+import { mapGetters } from 'vuex'
+import { productsInCart, closeMicrocart, isMicrocartOpen, removeFromCart } from 'core/api/cart'
 import onEscapePress from 'core/mixins/onEscapePress'
 
 export default {
@@ -35,9 +36,9 @@ export default {
     onNetworkStatusChanged (status) {
       this.isOnline = status.online
     },
-    closeMicrocart () {
+    closeMicrocartExtend () {
+      this.closeMicrocart()
       this.$store.commit('ui/setSidebar', false)
-      this.$store.commit('ui/setMicrocart', false)
       this.addCouponPressed = false
     },
     removeCoupon () {
@@ -56,8 +57,7 @@ export default {
       if (e.keyCode === 13) {
         this.applyCoupon()
       }
-    },
-    ...mapActions({ 'removeFromCart': 'cart/removeItem' })
+    }
   },
   computed: {
     ...mapGetters({
@@ -69,15 +69,9 @@ export default {
     payment () {
       return this.$store.state.cart.payment
     },
-    items () {
-      return this.$store.state.cart.cartItems
-    },
     coupon () {
       return this.$store.state.cart.platformTotals && this.$store.state.cart.platformTotals.hasOwnProperty('coupon_code') ? this.$store.state.cart.platformTotals.coupon_code : ''
-    },
-    ...mapState({
-      isOpen: state => state.ui.microcart
-    })
+    }
   },
-  mixins: [onEscapePress]
+  mixins: [onEscapePress, productsInCart, isMicrocartOpen, closeMicrocart, removeFromCart]
 }
