@@ -13,7 +13,8 @@ export default {
         region: '',
         country: '',
         postcode: '',
-        taxId: ''
+        taxId: '',
+        phone: ''
       },
       countries: Countries,
       changePassword: false,
@@ -54,7 +55,7 @@ export default {
     edit () {
       this.isEdited = true
     },
-    objectsEqual (a, b) {
+    objectsEqual (a, b, excludedFields = []) {
       const aProps = Object.keys(a)
       const bProps = Object.keys(b)
 
@@ -71,7 +72,7 @@ export default {
             if (!this.objectsEqual(a[propName], b[propName])) {
               return false
             }
-          } else if (a[propName] !== b[propName]) {
+          } else if (!excludedFields.includes(propName) && a[propName] !== b[propName]) {
             return false
           }
         }
@@ -80,7 +81,7 @@ export default {
     },
     updateProfile () {
       let updatedProfile
-      if (!this.objectsEqual(this.currentUser, this.$store.state.user.current) ||
+      if (!this.objectsEqual(this.currentUser, this.$store.state.user.current, ['updated_at', 'addresses']) ||
         !this.objectsEqual(this.userCompany, this.getUserCompany()) ||
         (this.userCompany.company && !this.addCompany)
       ) {
@@ -108,6 +109,7 @@ export default {
               updatedProfile.addresses[index].country_id = this.userCompany.country || ''
               updatedProfile.addresses[index].postcode = this.userCompany.postcode || ''
               updatedProfile.addresses[index].vat_id = this.userCompany.taxId || ''
+              updatedProfile.addresses[index].telephone = this.userCompany.phone || ''
             } else {
               updatedProfile.addresses.splice(index, 1)
               this.userCompany = {
@@ -118,7 +120,8 @@ export default {
                 region: '',
                 country: '',
                 postcode: '',
-                taxId: ''
+                taxId: '',
+                phone: ''
               }
             }
           }
@@ -133,6 +136,7 @@ export default {
             country_id: this.userCompany.country,
             postcode: this.userCompany.postcode,
             vat_id: this.userCompany.taxId,
+            telephone: this.userCompany.phone,
             default_billing: true
           })
         }
@@ -148,7 +152,7 @@ export default {
     exitSection (event, updatedProfile) {
       this.$bus.$emit('myAccount-before-updateUser', updatedProfile)
       if (!updatedProfile) {
-        this.currentUser = this.$store.state.user.current
+        this.currentUser = Object.assign({}, this.$store.state.user.current)
         this.userCompany = this.getUserCompany()
         this.changePassword = false
         this.oldPassword = ''
@@ -164,23 +168,25 @@ export default {
       }
     },
     getUserCompany () {
-      if (this.currentUser.hasOwnProperty('default_billing')) {
+      let user = this.$store.state.user.current
+      if (user.hasOwnProperty('default_billing')) {
         let index
         for (let i = 0; i < this.currentUser.addresses.length; i++) {
-          if (this.currentUser.addresses[i].id === Number(this.currentUser.default_billing)) {
+          if (user.addresses[i].id === Number(user.default_billing)) {
             index = i
           }
         }
         if (index >= 0) {
           return {
-            company: this.currentUser.addresses[index].company || '',
-            street: this.currentUser.addresses[index].street[0] || '',
-            house: this.currentUser.addresses[index].street[1] || '',
-            city: this.currentUser.addresses[index].city || '',
-            region: this.currentUser.addresses[index].region.region ? this.currentUser.addresses[index].region.region : '',
-            country: this.currentUser.addresses[index].country_id || '',
-            postcode: this.currentUser.addresses[index].postcode || '',
-            taxId: this.currentUser.addresses[index].vat_id || ''
+            company: user.addresses[index].company || '',
+            street: user.addresses[index].street[0] || '',
+            house: user.addresses[index].street[1] || '',
+            city: user.addresses[index].city || '',
+            region: user.addresses[index].region.region ? user.addresses[index].region.region : '',
+            country: user.addresses[index].country_id || '',
+            postcode: user.addresses[index].postcode || '',
+            taxId: user.addresses[index].vat_id || '',
+            phone: user.addresses[index].telephone || ''
           }
         }
       } else {
@@ -192,7 +198,8 @@ export default {
           region: '',
           country: '',
           postcode: '',
-          taxId: ''
+          taxId: '',
+          phone: ''
         }
       }
     },
