@@ -1,11 +1,11 @@
-import moment from 'moment'
-import config from 'config'
+import { currentStoreView } from '@vue-storefront/store/lib/multistore'
 
 /**
  * Converts date to format provided as an argument or defined in config file (if argument not provided)
  * @param {String} date
  * @param {String} format
  */
-export function date (date, format = config.i18n.dateFormat) {
-  return moment(date).format(format)
+export function date (date, format) {
+  const storeView = currentStoreView()
+  return new Date(date).toLocaleString(storeView.i18n.defaultLocale)
 }
