@@ -3,25 +3,6 @@ import Vuex from 'vuex'
 import * as types from './mutation-types'
 import localForage from 'localforage'
 import UniversalStorage from './lib/storage'
-import order from './modules/order'
-import product from './modules/product'
-import category from './modules/category'
-import attribute from './modules/attribute'
-import cart from './modules/cart'
-import wishlist from './modules/wishlist'
-import compare from './modules/compare'
-import user from './modules/user'
-import payment from './modules/payment'
-import shipping from './modules/shipping'
-import ui from './modules/ui-store'
-import checkout from './modules/checkout'
-import homepage from './modules/homepage'
-import stock from './modules/stock'
-import tax from './modules/tax'
-import social from './modules/social-tiles'
-import claims from './modules/claims'
-import sync from './modules/sync'
-import promoted from './modules/promoted-offers'
 import { currentStoreView } from './lib/multistore'
 
 if (!global.$VS) global.$VS = {}
@@ -56,7 +37,7 @@ const plugins = [
         console.error('Store mutation name is incorrectly formed')
       }
 
-      if (actionName === types.CART_ADD_ITEM || actionName === types.CART_DEL_ITEM || actionName === types.CART_UPD_ITEM || actionName === types.CART_UPD_ITEM_PROPS) { // check if this mutation is cart related
+      if (actionName === types.CART_LOAD_CART || actionName === types.CART_ADD_ITEM || actionName === types.CART_DEL_ITEM || actionName === types.CART_UPD_ITEM || actionName === types.CART_UPD_ITEM_PROPS) { // check if this mutation is cart related
         global.$VS.db.cartsCollection.setItem('current-cart', state.cart.cartItems).catch((reason) => {
           console.error(reason) // it doesn't work on SSR
         }) // populate cache
@@ -116,32 +97,17 @@ const plugins = [
           console.error(reason)
         })
       }
+      if (actionName === 'setCmsBlock' || actionName === 'setCmsPage') {
+        global.$VS.db.cmsData.setItem('cms-data', state.cms).catch((reason) => {
+          console.error(reason)
+        })
+      }
     })
   }
 ]
 
 let rootStore = new Vuex.Store({
-  modules: { // TODO: refactor it to return just the constructor to avoid event-bus and i18n shenigans; challenge: the singleton management OR add i18n and eventBus here to rootStore instance?  modules: {
-    order,
-    product,
-    category,
-    attribute,
-    cart,
-    wishlist,
-    compare,
-    user,
-    payment,
-    shipping,
-    ui,
-    homepage,
-    social,
-    stock,
-    checkout,
-    tax,
-    claims,
-    sync,
-    promoted
-  },
+  // TODO: refactor it to return just the constructor to avoid event-bus and i18n shenigans; challenge: the singleton management OR add i18n and eventBus here to rootStore instance?  modules: {
   state,
   mutations,
   plugins
@@ -182,59 +148,77 @@ rootStore.init = function (config, i18n = null, eventBus = null) { // TODO: init
   Vue.prototype.$db = {
     ordersCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'orders'
+      storeName: 'orders',
+      driver: localForage[config.localForage.defaultDrivers['orders']]
     })),
     categoriesCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'categories'
+      storeName: 'categories',
+      driver: localForage[config.localForage.defaultDrivers['categories']]
     })),
     attributesCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'attributes'
+      storeName: 'attributes',
+      driver: localForage[config.localForage.defaultDrivers['attributes']]
     })),
     cartsCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'carts'
+      storeName: 'carts',
+      driver: localForage[config.localForage.defaultDrivers['carts']]
     })),
     elasticCacheCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'elasticCache'
+      storeName: 'elasticCache',
+      driver: localForage[config.localForage.defaultDrivers['elasticCache']]
     })),
     productsCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'products'
+      storeName: 'products',
+      driver: localForage[config.localForage.defaultDrivers['products']]
     })),
     claimsCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'claims'
+      storeName: 'claims',
+      driver: localForage[config.localForage.defaultDrivers['claims']]
     })),
     wishlistCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'wishlist'
+      storeName: 'wishlist',
+      driver: localForage[config.localForage.defaultDrivers['wishlist']]
     })),
     compareCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'compare'
+      storeName: 'compare',
+      driver: localForage[config.localForage.defaultDrivers['compare']]
     })),
     usersCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'user'
+      storeName: 'user',
+      driver: localForage[config.localForage.defaultDrivers['user']]
     })),
     syncTaskCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'syncTasks'
+      storeName: 'syncTasks',
+      driver: localForage[config.localForage.defaultDrivers['syncTasks']]
     })),
     checkoutFieldsCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'checkoutFieldValues'
+      storeName: 'checkoutFieldValues',
+      driver: localForage[config.localForage.defaultDrivers['checkoutFieldValues']]
     })),
     newsletterPreferencesCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'newsletterPreferences'
+      storeName: 'newsletterPreferences',
+      driver: localForage[config.localForage.defaultDrivers['newsletterPreferences']]
     })),
     ordersHistoryCollection: new UniversalStorage(localForage.createInstance({
       name: dbNamePrefix + 'shop',
-      storeName: 'ordersHistory'
+      storeName: 'ordersHistory',
+      driver: localForage[config.localForage.defaultDrivers['ordersHistory']]
+    })),
+    cmsData: new UniversalStorage(localForage.createInstance({
+      name: dbNamePrefix + 'shop',
+      storeName: 'cms'
     }))
   }
   global.$VS.db = Vue.prototype.$db // localForage instance
