@@ -2,6 +2,7 @@ import * as types from '../../mutation-types'
 import { slugify, breadCrumbRoutes } from '../../helpers'
 import { entityKeyName } from '../../lib/entities'
 import EventBus from '../../lib/event-bus'
+import config from 'config'
 
 export default {
   [types.CATEGORY_UPD_CURRENT_CATEGORY] (state, category) {
@@ -19,7 +20,7 @@ export default {
     for (let category of state.list) {
       let catSlugSetter = (category) => {
         for (let subcat of category.children_data) { // TODO: fixme and move slug setting to vue-storefront-api
-          subcat = Object.assign(subcat, { slug: subcat.hasOwnProperty('name') ? slugify(subcat.name) + '-' + subcat.id : '' })
+          subcat = Object.assign(subcat, { slug: (subcat.hasOwnProperty('url_key') && config.products.useMagentoUrlKeys) ? subcat.url_key : (subcat.hasOwnProperty('name') ? slugify(subcat.name) + '-' + subcat.id : '') })
           catSlugSetter(subcat)
         }
       }
