@@ -8,5 +8,6 @@ export default [
   require('@vue-storefront/extension-payment-backend-methods/index.js'),
   require('@vue-storefront/extension-payment-cash-on-delivery/index.js'),
   require('@vue-storefront/extension-template/index.js'),
-  require('vsf-payment-stripe/index.js')
+  require('vsf-payment-stripe/index.js'),
+  require('src/extensions/cms/index.js')
 ]
