@@ -12,7 +12,8 @@ export default function (app, router, store, config) {
   store.registerModule(EXTENSION_KEY, extensionStore) // add custom store
   // TODO: register module events here
   app.$on('application-after-init', () => {
-    console.log('custom-event')
+    console.log('Example custom-event added by template example')
+    store.dispatch('product/example')
   })
 
   EventBus.$on('thankyoupage-after-mounted', (parent) => {
