@@ -11,11 +11,13 @@
 
 <script>
 import NoSSR from 'vue-no-ssr'
+import VueOfflineMixin from 'vue-offline/mixin'
 
 export default {
   components: {
     'no-ssr': NoSSR
-  }
+  },
+  mixins: [VueOfflineMixin]
 }
 </script>
 <style scoped>
