@@ -18,6 +18,7 @@
           name="email"
           focus
           v-model="email"
+          @blur="$v.email.$touch()"
           :placeholder="$t('E-mail address *')"
           :validations="[
             {
@@ -35,6 +36,7 @@
           type="password"
           name="password"
           v-model="password"
+          @blur="$v.password.$touch()"
           :placeholder="$t('Password *')"
           :validation="{
             condition: !$v.password.required && $v.password.$error,
@@ -80,13 +82,7 @@ import { required, email } from 'vuelidate/lib/validators'
 import i18n from 'core/lib/i18n'
 
 export default {
-  data () {
-    return {
-      remember: false,
-      email: '',
-      password: ''
-    }
-  },
+  mixins: [Login],
   validations: {
     email: {
       required,
@@ -96,24 +92,9 @@ export default {
       required
     }
   },
-  mixins: [Login],
   methods: {
-    switchElem () {
-      this.$store.commit('ui/setAuthElem', 'register')
-    },
     close () {
       this.$bus.$emit('modal-hide', 'modal-signup')
-    },
-    remindPassword () {
-      if (!(typeof navigator !== 'undefined' && navigator.onLine)) {
-        this.$bus.$emit('notification', {
-          type: 'error',
-          message: i18n.t('Reset password feature does not work while offline!'),
-          action1: { label: i18n.t('OK'), action: 'close' }
-        })
-      } else {
-        this.$store.commit('ui/setAuthElem', 'forgot-pass')
-      }
     },
     login () {
       if (this.$v.$invalid) {
