@@ -4,8 +4,8 @@
       <h3 class="cl-accent ml30 mt50 summary-title">
         {{ $t('Order Summary') }}
       </h3>
-      <product v-for="product in items" :key="product.sku" :product="product"/>
-      <div v-if="items.length" class="checkout bg-cl-secondary pt10 serif cl-accent">
+      <product v-for="product in productsInCart" :key="product.sku" :product="product"/>
+      <div v-if="productsInCart && productsInCart.length" class="checkout bg-cl-secondary pt10 serif cl-accent">
 
         <div v-for="(segment, index) in totals" :key="index" class="row pt15 pb20 pl30 pr55 " v-if="segment.code !== 'grand_total'">
           <div class="col-xs cl-accent">
