@@ -1,19 +1,19 @@
 <template>
   <div
     class="microcart mw-100 fixed cl-accent"
-    :class="[items.length ? 'bg-cl-secondary' : 'bg-cl-primary', { active: isOpen }]"
+    :class="[productsInCart.length ? 'bg-cl-secondary' : 'bg-cl-primary', { active: isMicrocartOpen }]"
   >
     <div class="row middle-xs bg-cl-primary top-sm">
       <div class="col-xs-10">
         <h2
-          v-if="items.length"
+          v-if="productsInCart.length"
           class="cl-accent mt60 mb35 ml40 heading"
         >
           {{ $t('Shopping cart') }}
         </h2>
       </div>
       <div class="col-xs-2 end-xs">
-        <button type="button" class="p0 brdr-none bg-cl-transparent close" @click="closeMicrocart">
+        <button type="button" class="p0 brdr-none bg-cl-transparent close" @click="closeMicrocartExtend">
           <i class="material-icons p15 cl-accent">
             close
           </i>
@@ -21,20 +21,20 @@
       </div>
     </div>
 
-    <h4 v-if="!items.length" class="cl-accent ml30">
+    <h4 v-if="!productsInCart.length" class="cl-accent ml30">
       {{ $t('Your shopping cart is empty.') }}
     </h4>
-    <div v-if="!items.length" class="ml30" @click="closeMicrocart">
+    <div v-if="!productsInCart.length" class="ml30" @click="closeMicrocartExtend">
       {{ $t("Don't hesitate and") }}
       <router-link :to="localizedRoute('/')">
         {{ $t('browse our catalog') }}
       </router-link>
       {{ $t('to find something beautiful for You!') }}
     </div>
-    <ul v-if="items.length" class="bg-cl-primary m0 px40 pb40 products">
-      <product v-for="product in items" :key="product.sku" :product="product" />
+    <ul v-if="productsInCart.length" class="bg-cl-primary m0 px40 pb40 products">
+      <product v-for="product in productsInCart" :key="product.sku" :product="product" />
     </ul>
-    <div v-if="items.length" class="summary px40 cl-accent serif">
+    <div v-if="productsInCart.length" class="summary px40 cl-accent serif">
       <h3 class="m0 pt40 mb30 weight-400 summary-heading">
         {{ $t('Shopping summary') }}
       </h3>
@@ -81,11 +81,11 @@
     </div>
     <div
       class="row py20 px40 middle-xs actions"
-      v-if="items.length && !isCheckoutMode"
+      v-if="productsInCart.length && !isCheckoutMode"
     >
       <div class="col-xs-12 col-sm first-sm">
         <router-link :to="localizedRoute('/')" class="no-underline cl-secondary link">
-          <span @click="closeMicrocart">
+          <span @click="closeMicrocartExtend">
             {{ $t('Return to shopping') }}
           </span>
         </router-link>
@@ -93,7 +93,7 @@
       <div class="col-xs-12 first-xs col-sm-4 end-sm">
         <button-full
           :link="{ name: 'checkout' }"
-          @click.native="closeMicrocart"
+          @click.native="closeMicrocartExtend"
         >
           {{ $t('Go to checkout') }}
         </button-full>
