@@ -1,8 +1,8 @@
 <template>
-  <section v-if="!singleBanner" class="offers container my30">
-    <div class="row m0">
+  <section v-if="!singleBanner" class="offers container my30 px15">
+    <div class="row">
       <div
-        class="offer-container col-xs-12 col-sm-6 px15"
+        class="offer-container col-xs-12 col-sm-6 pb15"
         v-for="(banner, index) in banners.mainBanners"
         :key="index"
       >
@@ -21,9 +21,9 @@
         </router-link>
       </div>
 
-      <div class="col-xs-12 col-sm-6 p0">
+      <div class="col-xs-12 col-sm-6">
         <div
-          class="offer-container px15 pb30"
+          class="offer-container pb15"
           v-for="(banner, index) in banners.smallBanners"
           :key="index"
         >
@@ -40,25 +40,27 @@
       </div>
     </div>
   </section>
-  <section v-else class="container my30">
-    <div
-      class="col-xs-12 px15"
-      v-for="(banner, index) in banners.productBanners"
-      :key="index"
-    >
-      <router-link :to="localizedRoute(banner.link)">
-        <div
-          class="offer offer-product border-box p5 flex center-xs middle-xs cl-white bg-cl-th-accent"
-          v-lazy:background-image="banner.image"
-        >
-          <p class="subtitle m0 serif h3 uppercase">
-            {{ banner.subtitle }}
-          </p>
-          <h2 class="title m0 h1">
-            {{ banner.title }}
-          </h2>
-        </div>
-      </router-link>
+  <section v-else class="container my30 px15">
+    <div class="row">
+      <div
+        class="col-xs-12"
+        v-for="(banner, index) in banners.productBanners"
+        :key="index"
+      >
+        <router-link :to="localizedRoute(banner.link)">
+          <div
+            class="offer offer-product border-box p5 flex center-xs middle-xs cl-white bg-cl-th-accent"
+            v-lazy:background-image="banner.image"
+          >
+            <p class="subtitle m0 serif h3 uppercase">
+              {{ banner.subtitle }}
+            </p>
+            <h2 class="title m0 h1">
+              {{ banner.title }}
+            </h2>
+          </div>
+        </router-link>
+      </div>
     </div>
   </section>
 </template>
@@ -93,22 +95,13 @@ export default {
 </script>
 
 <style lang="scss" scoped>
- .offers {
-    @media (max-width: 767px) {
-      padding: 0 15px;
-    }
-  }
   .offer-container {
     &:last-child {
       padding-bottom: 0;
     }
-
-    @media (max-width: 767px) {
-      padding: 5px 0;
-    }
   }
   .offer {
-    height: 750px;
+    height: 735px;
     flex-direction: column;
     background-position: center;
     background-size: cover;
