@@ -3,7 +3,10 @@
     <header class="bg-cl-secondary py35 pl20">
       <div class="container">
         <breadcrumbs :routes="breadcrumbs.routes" :active-route="category.name" />
-        <h1 class="category-title mb10"> {{ category.name }} </h1>
+        <div class="row middle-sm">
+          <h1 class="col-sm-9 category-title mb10"> {{ category.name }} </h1>
+          <div class="sorting col-sm-3 align-right"><SortBy /></div>
+        </div>
       </div>
       <div class="container">
         <div class="row m0">
@@ -13,11 +16,12 @@
           >
             {{ $t('Filters') }}
           </button>
+          <div class="mobile-sorting col-xs-6 mt25"><SortBy /></div>
         </div>
       </div>
     </header>
     <div class="container pb60">
-      <div class="row m0 pt15 center-md">
+      <div class="row m0 pt15">
         <div class="col-md-3 start-xs category-filters">
           <sidebar :filters="filters.available"/>
         </div>
@@ -45,13 +49,15 @@ import Category from 'core/pages/Category'
 import Sidebar from '../components/core/blocks/Category/Sidebar.vue'
 import ProductListing from '../components/core/ProductListing.vue'
 import Breadcrumbs from '../components/core/Breadcrumbs.vue'
+import SortBy from '../components/core/SortBy.vue'
 // import builder from 'bodybuilder'
 
 export default {
   components: {
     ProductListing,
     Breadcrumbs,
-    Sidebar
+    Sidebar,
+    SortBy
   },
   data () {
     return {
@@ -93,6 +99,10 @@ export default {
     display: none;
   }
 
+  .mobile-sorting {
+    display: none;
+  }
+
   .category-title {
     line-height: 65px;
   }
@@ -122,6 +132,14 @@ export default {
     .mobile-filters-button {
       display: block;
       height: 45px;
+    }
+
+    .sorting {
+      display: none;
+    }
+
+    .mobile-sorting {
+      display: block;
     }
 
     .category-filters {
