@@ -1,7 +1,6 @@
-// Add your own plugins in this folder by exporting object with their instances
-// They will be registered later with Vue.use(). You can find examples in core/plugins/index.js
+// Add your own plugins in this folder with Vue.use()
+import Vue from 'vue'
 
-// Export plugin objects here. They will be proceeded by Vue.use()
-export {
+import VueProgressBar from 'vue-progressbar'
 
-}
+Vue.use(VueProgressBar)
