@@ -13,16 +13,17 @@ export default {
     { name: 'apple-mobile-web-app-status-bar-style', content: '#ffffff' }
   ],
   link: [
-    { rel: 'icon', href: '/assets/logo.png' },
-    { rel: 'apple-touch-icon', href: '/assets/launcher-icon-1x.png' },
-    { rel: 'apple-touch-startup-image', href: '/assets/launcher-icon-144x144.png' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_2048.png', sizes: '2048x2732' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_1668.png', sizes: '1668x2224' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_1536.png', sizes: '1536x2048' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_1125.png', sizes: '1125x2436' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_1242.png', sizes: '1242x2208' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_750.png', sizes: '750x1334' },
-    { rel: 'apple-touch-icon', href: '/assets/apple_splash_640.png', sizes: '640x1136' },
-    { rel: 'manifest', href: '/assets/manifest.json' }
+    { rel: 'icon', type: 'image/png', href: '/assets/favicon-32x32.png', sizes: '32x32' },
+    { rel: 'icon', type: 'image/png', href: '/assets/favicon-16x16.png', sizes: '16x16' },
+    { rel: 'apple-touch-icon', href: '/assets/apple-touch-icon.png' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_2048.png', sizes: '2048x2732' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_1668.png', sizes: '1668x2224' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_1536.png', sizes: '1536x2048' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_1125.png', sizes: '1125x2436' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_1242.png', sizes: '1242x2208' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_750.png', sizes: '750x1334' },
+    { rel: 'apple-touch-startup-image', href: '/assets/apple_splash_640.png', sizes: '640x1136' },
+    { rel: 'manifest', href: '/assets/manifest.json' },
+    { rel: 'stylesheet', href: 'https://fonts.googleapis.com/icon?family=Material+Icons' }
   ]
 }
