@@ -1,7 +1,7 @@
 <template>
   <div class="microcart">
     <h3>Microcart</h3>
-    <div v-for="(item, index) in items" :key="index"> {{ item.name }} ({{ item.qty }}) </div>
+    <div v-for="(item, index) in productsInCart" :key="index"> {{ item.name }} ({{ item.qty }}) </div>
   </div>
 </template>
 
