@@ -1,7 +1,3 @@
-// Add your own plugins in this folder by exporting object with their instances
-// They will be registered later with Vue.use(). You can find examples in core/plugins/index.js
-
-// Export plugin objects here. They will be proceeded by Vue.use()
-export {
-
-}
+// import Vue from 'vue'
+// uncommet import if you want to add some plugins
+// Add your plugins with Vue.use()
