@@ -1,6 +1,7 @@
 // Here you can place theme-specific stores that will be added to the core stores.
 // Everything that you will export here will be treated as a Vuex module
 // The good practice is to keep all modulase separated in this folder.
+// You can also override core stores here
 
 // Below you can find commented exaple for additional theme store
 // const exampleThemeStore = {
@@ -14,5 +15,5 @@
 // }
 
 export default {
-  // exampleThemeStore
+  // exampel store
 }
