@@ -1,3 +1,5 @@
-// require all test files (files that ends with .spec.js)
-var testsContext = require.context('./specs', true, /\.spec$/)
-testsContext.keys().forEach(testsContext)
+
+//  Core API modules tests
+var coreApiTest = require.context('../../core/api', true, /\.spec$/)
+// testsContext.keys().forEach(testsContext)
+coreApiTest.keys().forEach(coreApiTest)
