@@ -1,11 +1,14 @@
-const baseConfig = require('../../core/build/webpack.base.config')
+const baseConfig = require('../../core/build/webpack.base.config.js')
 const webpack = require('webpack')
+const VueLoaderPlugin = require('vue-loader/lib/plugin')
+
 const webpackConfig = Object.assign({}, baseConfig, {
   devtool: 'inline-source-map',
   plugins: [
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': '"test"'
-    })
+    }),
+    new VueLoaderPlugin()
   ]
 })
 
@@ -14,7 +17,7 @@ delete webpackConfig.entry
 
 module.exports = function (config) {
   config.set({
-    browsers: ['PhantomJS'],
+    browsers: ['Chrome'],
     frameworks: ['mocha', 'sinon-chai'],
     reporters: ['spec'],
     files: ['./index.js'],
